bot/pyproject.toml
[project]
name = "cukiller"
version = "0.1.0"
description = "Бот для игры в Killer в Telegram для Центрального Университета"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.22.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.11.0",
    "tortoise-orm>=0.25.1",
    "pillow>=12.0.0",
    "aiogram-dialog==2.4.0",
    "prometheus-client>=0.21.0",
    "requests==2.32.5",
    "redis==7.0.1",
    "aerich==0.9.2",
    "ruff>=0.14.10",
]

[tool.ruff]
builtins = []
cache-dir = ".ruff_cache"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "dist",
    "migrations",
    "node_modules",
    "venv",
    "manage.py"
]
extend-exclude = []
extend-include = []
fix = false
fix-only = false
force-exclude = true
include = ["*.py", "*.pyi", "*.ipynb", "**/pyproject.toml"]
indent-width = 4
line-length = 120

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 120
exclude = []
indent-style = "space"
line-ending = "lf"
preview = false
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint]
exclude = ["tests.py"]
fixable = ["ALL"]
ignore = [
    "ARG",
    "COM812",
    "D",
    "ISC001",
    "PT009",
    "N813",
    "RUF012",
    "ANN201",
    "RUF001",
]
select = ["ALL"]
task-tags = ["TODO"]

[tool.aerich]
tortoise_orm = "services.settings.settings.tortoise_config"
location = "./migrations"
src_folder = "./."
